+++ app/src/main/java/com/tosslab/jandi/app/local/orm/repositories/SendMessageRepository.java
@@ -68,7 +68,7 @@ public class SendMessageRepository {
 
     }
 
-    public List<SendMessage> getSendMessage(int roomId) {
+    public List<SendMessage> getSendMessage(long roomId) {
         try {
             Dao<SendMessage, ?> dao = helper.getDao(SendMessage.class);
             return dao.queryBuilder()
@@ -141,7 +141,7 @@ public class SendMessageRepository {
 
     }
 
-    public int updateSendMessageStatus(long localId, int messageId, SendMessage.Status status) {
+    public int updateSendMessageStatus(long localId, long messageId, SendMessage.Status status) {
         try {
             Dao<SendMessage, ?> dao = helper.getDao(SendMessage.class);
             UpdateBuilder<SendMessage, ?> updateBuilder = dao.updateBuilder();
@@ -156,14 +156,14 @@ public class SendMessageRepository {
         return 0;
     }
 
-    public void deleteCompletedMessages(List<Integer> messageIds) {
+    public void deleteCompletedMessages(List<Long> messageIds) {
         try {
             Dao<SendMessage, ?> dao = helper.getDao(SendMessage.class);
 
 
             dao.callBatchTasks(() -> {
 
-                for (Integer messageId : messageIds) {
+                for (long messageId : messageIds) {
                     DeleteBuilder<SendMessage, ?> deleteBuilder = dao.deleteBuilder();
                     deleteBuilder.where().eq("messageId", messageId);
                     deleteBuilder.delete();
