+++ app/src/main/java/com/tosslab/jandi/app/network/client/main/IMainRestApiAuth.java
@@ -19,9 +19,9 @@ public interface IMainRestApiAuth {
 
     ResAccountInfo updatePrimaryEmailByMainRest(ReqUpdatePrimaryEmailInfo updatePrimaryEmailInfo) throws RetrofitError;
 
-    ResLeftSideMenu getInfosForSideMenuByMainRest(int teamId) throws RetrofitError;
+    ResLeftSideMenu getInfosForSideMenuByMainRest(long teamId) throws RetrofitError;
 
-    ResCommon setMarkerByMainRest(int entityId, ReqSetMarker reqSetMarker) throws RetrofitError;
+    ResCommon setMarkerByMainRest(long entityId, ReqSetMarker reqSetMarker) throws RetrofitError;
 
     ResSearchFile searchFileByMainRest(ReqSearchFile reqSearchFile) throws RetrofitError;
