+++ app/src/main/java/com/tosslab/jandi/app/network/models/ResUpdateMessages.java
@@ -11,7 +11,7 @@ import java.util.List;
 @JsonIgnoreProperties(ignoreUnknown = true)
 @JsonSerialize(include = JsonSerialize.Inclusion.NON_NULL)
 public class ResUpdateMessages {
-    public int lastLinkId;
+    public long lastLinkId;
     public int messageCount;
     public List<ResMessages.Link> messages;
