+++ app/src/main/java/com/tosslab/jandi/app/network/models/ResUploadedFile.java
@@ -11,23 +11,23 @@ import org.codehaus.jackson.map.annotate.JsonSerialize;
 @JsonSerialize(include = JsonSerialize.Inclusion.NON_NULL)
 public class ResUploadedFile {
 
-    private int messageId;
-    private int commentId;
+    private long messageId;
+    private long commentId;
     private UploadInfo fileInfo;
 
-    public int getMessageId() {
+    public long getMessageId() {
         return messageId;
     }
 
-    public void setMessageId(int messageId) {
+    public void setMessageId(long messageId) {
         this.messageId = messageId;
     }
 
-    public int getCommentId() {
+    public long getCommentId() {
         return commentId;
     }
 
-    public void setCommentId(int commentId) {
+    public void setCommentId(long commentId) {
         this.commentId = commentId;
     }
